+++ components/contact-form.tsx
@@ -1,7 +1,13 @@
 "use client"
 
 import { QuizModal } from "./quiz-modal"
+import { SimpleContactFormModal } from "./simple-contact-form-modal"
 
 export function ContactForm() {
-  return <QuizModal />
+  return (
+    <>
+      <QuizModal />
+      <SimpleContactFormModal />
+    </>
+  )
 }